rename InsightsReader
