+++ internal/saucecloud/jobservice.go
@@ -90,8 +90,8 @@ func (s JobService) StartJob(ctx context.Context, opts job.StartOptions) (jobID 
 }
 
 type JobCommandService struct {
-	InsightsReader insights.Service
-	UserService    iam.Service
+	Commander   insights.Service
+	UserService iam.Service
 }
 
 func (s JobCommandService) ListJobs(ctx context.Context, jobSource string, queryOption job.QueryOption) (job.List, error) {
@@ -99,9 +99,9 @@ func (s JobCommandService) ListJobs(ctx context.Context, jobSource string, query
 	if err != nil {
 		return job.List{}, fmt.Errorf("failed to get user: %w", err)
 	}
-	return s.InsightsReader.ListJobs(ctx, user.ID, jobSource, queryOption)
+	return s.Commander.ListJobs(ctx, user.ID, jobSource, queryOption)
 }
 
 func (s JobCommandService) ReadJob(ctx context.Context, id string) (job.Job, error) {
-	return s.InsightsReader.ReadJob(ctx, id)
+	return s.Commander.ReadJob(ctx, id)
 }
